move //logic/transfomer/mapping to //reasoner/resolution/answer
