+++ reasoner/resolution/answer/Mapping.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.logic.transformer;
+package grakn.core.reasoner.resolution.answer;
 
 import grakn.core.concept.Concept;
 import grakn.core.concept.answer.ConceptMap;
