+++ reasoner/resolution/resolver/RootResolver.java
@@ -22,7 +22,7 @@ import grakn.common.collection.Pair;
 import grakn.common.concurrent.actor.Actor;
 import grakn.core.concept.answer.ConceptMap;
 import grakn.core.logic.resolvable.Concludable;
-import grakn.core.logic.transformer.Mapping;
+import grakn.core.reasoner.resolution.answer.Mapping;
 import grakn.core.pattern.Conjunction;
 import grakn.core.reasoner.resolution.MockTransaction;
 import grakn.core.reasoner.resolution.ResolutionRecorder;
