+++ reasoner/resolution/resolver/RuleResolver.java
@@ -23,7 +23,7 @@ import grakn.common.concurrent.actor.Actor;
 import grakn.core.concept.answer.ConceptMap;
 import grakn.core.logic.Rule;
 import grakn.core.logic.resolvable.Concludable;
-import grakn.core.logic.transformer.Mapping;
+import grakn.core.reasoner.resolution.answer.Mapping;
 import grakn.core.pattern.Conjunction;
 import grakn.core.reasoner.resolution.MockTransaction;
 import grakn.core.reasoner.resolution.ResolverRegistry;
